Methods and types

Data entry

The methods below are used to enter data into Grumps.

Grumps.Sources — Method

Sources( 
    T           = DefaultSourceTypes; 
    consumers   :: Any = nothing, 
    products    :: Any = nothing, 
    marketsizes :: Any = nothing, 
    draws       :: Any = nothing,
    user        :: Any = nothing
)

Creates a GrumpsSources object with source type entries of type T where the entries are provided in the optional parameters.

Grumps (potentially) uses four data sources: a data source for consumer-level data, one for product-level data, one for market size information, and one for demographic draws. Only the product-level data are required, but are by themselves insufficient. For instance, for BLP95 one needs information on products, market sizes, and demographics; for the Grumps estimator one needs all four types of data; for a multinomial logit both consumer and product information are needed. Not all data are needed for all markets. For instance, it is ok for some estimators for there to be consumer-level data in some markets but not others.

The T argument is mostly there to allow for future expansion, so the description below applies to the case in which T = DefaultSourceTypes.

By default, the entries can be nothing, a string, a DataFrame, or a SourceFileType. If an entry is nothing, it means that no such data is to be used. If an entry is a string then it is converted to a SourceFileCSV entry with comma delimiter where the string name is the file name. To use other source file types, create a SourceFileType first. A DataFrame can be passed, also. In all cases other than nothing, data will eventually be (converted to) a DataFrame and parsed from that.

The consumers variable specifies where consumer-level data can be found, the products variable is for the product-level data, marketsizes is for market sizes, and draws is for demographic draws; user has not been implemented yet.

Use the Variables method to specify the way the data sources are formatted and the specification to estimate.

Grumps.Variables — Method

function Variables( ; 
  market              :: Symbol = :market,
  choice              :: Symbol = :choice,
  interactions        :: Mat{Symbol} = [],
  randomcoefficients  :: Vec{Symbol} = [],
  outsidegood         :: String = "outsidegood",
  share               :: Symbol = :share,
  marketsize          :: Symbol = :N,
  regressors          :: Vec{Symbol} = [],
  instruments         :: Vec{Symbol} = [],
  dummies             :: Vec{Symbol} = [],
  nuisancedummy       :: Symbol = :none,
  microinstruments    :: Mat{Symbol} = [],
  user                :: Mat{Symbol} = []
    )

This method creates an object of type GrumpsVariables. It contains references to the variables that Grumps uses to create variables from the data sources specified by the call to the Sources function.

For instance, market is the column heading in the source spreadsheets for the market indicator. This get's passed as a symbol, so the default (:market) says that the column heading is market, which is both case and spaces sensititve. The same column heading is used across all sources. All entries with the exception of outsidegood refer to the column heading: outsidegood refers to the label used for the outside good, which should be the same across both spreadsheets and markets.

market refers to the variable containing the market indicator in all input datasets

product refers to the variable containing the product indicator in the product dataset

choice refers to the variable indicating the choice indicator in the consumer level datasets

interactions refers to the variables indicating consumer and product variable interactions (each row contains consumer variable, product variable)

randomcoefficients refers to the product level variables that have a random coefficient on them

outsidegood refers to the label used for the outside good

share refers to the label used for the product level share

marketsize refers to the size of the market (number of people)

regressors refers to the label used for the second stage regressors

instruments refers to the label used for the second stage instruments

dummies refers to discrete variables to be converted to second stage dummy regressors and instruments

nuisancedummy refers to at most one variable to be converted to a second stage dummy regressors and instrument whose coefficient value is of no interest

microinstruments refers to micro instruments, which are only relevant for gmm style procedures

user refers to a list of variables to be added to the consumer-product interactions using a user-specified procedure

Optimization options

The default optimization options are sensible, in which case this section can be skipped. But for those who want to play with tolerances and such, have at it.

Grumps.OptimizationOptions — Method

OptimizationOptions(; 
θopt = OptimOptionsθ(), 
δopt = OptimOptionsδ(), 
threads = GrumpsThreads(), 
memsave = false, 
maxrepeats = 4, 
probtype = :fast )

Sets the options used for numerical optimization. θopt is used for the external optimization routine, δopt for the internal one. These are both of type OptimOptions; see the OptimOptionsθ and OptimOptionsδ methods for elaboration. The memsave variable is set to false by default; turning it on will reduce memory consumption significantly, but will also slow down computation. The variable maxrepeats may disappear in the future.

Finally, there are two ways of computing choice probabilities: robust and fast, specified by passing :robust or :fast in probtype. Fast choice probabilities are the default for good reason.

Grumps.OptimOptionsθ — Method

OptimOptionsθ(; 
f_tol = 1.0e-8, 
g_tol = 1.0e-4, 
x_tol = 1.0e-5, 
iterations = 25, 
show_trace = true, 
store_trace = true, 
extended_trace = true )

Creates and returns an OptimOptions optimization options variable for the outer optimization algorithm, including the function value tolerance, the gradient tolerance, the solution tolerance, the maximum number of iterations, whether to show the trace, whether to store the trace, and whether to keep the extended trace. See the Optim package for details.

The current version of Grumps will largely ignore the trace-related parameters.

Grumps.OptimOptionsδ — Method

OptimOptionsδ( ; 
f_tol = 1.0e-8, 
g_tol = 1.0e-8, 
x_tol = 1.0e-6, 
iterations = 25, 
show_trace = false, 
store_trace = true, 
extended_trace = false )

Creates and returns an OptimOptions optimization options variable for the inner optimization algorithm, including the function value tolerance, the gradient tolerance, the solution tolerance, the maximum number of iterations, whether to show the trace, whether to store the trace, and whether to keep the extended trace. See the Optim package for details.

The current version of Grumps will largely ignore the trace-related parameters.

Estimator choice

Grumps.Estimator — Method

Estimator( s :: String )

Creates and returns a GrumpsEstimator type. Grumps is reasonably good at figuring out what it is that you want, so e.g. Estimator( "maximum likelihood" ) gives you the unpenalized Grumps maximum likelihood estimator.

Grumps.Estimator — Method

Estimator( s :: Symbol )

Creates and returns a GrumpsEstimator type.

This is one method of specifying the estimator used. However, it unforgiving in that the exact symbol used internally must be passed, so the Estimator( s :: String ) method is usually a better choice.

Possible choices include:

:plm the penalized Grumps maximum likelihood estimator not implemented yet

:vanilla the unpenalized Grumps maximum likelihood estimator

:shareconstraint the unpenalized Grumps maximum likelihood estimator with share constraints

:gmm GMM estimator that uses both micro and macro moments

:mixedlogit mixed logit maximum likelihood estimator

Data object creation

to be done

Algorithm call

to be done

Retrieving results

As noted above, Grumps will return its results in a GrumpsSolution variable that can be queried as follows. to be expanded

Grumps.getθ — Method

getθ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for θ that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getδ — Method

getδ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for δ that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getβ — Method

getβ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for β that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getcoef — Method

getcoef( e :: GrumpsEstimate )

Returns the estimated coefficient value.

Grumps.getstde — Method

getstde( e :: GrumpsEstimate )

Returns the standard error.

Grumps.gettstat — Method

gettstat( e :: GrumpsEstimate )

Returns the t statistic.

Grumps.getname — Method

getname( e :: GrumpsEstimate )

Returns the variable name.

Extending Grumps