User Interface

The sections below describe the main calls needed to use Grumps. For any functions that are not documented here, simply use ? in the REPL, e.g. ?Variables.

The way that Grumps works is that one first specifies where the data are stored, what specification to use, which estimator to use, etcetera, before calling the functions that actually perform work with these choices. All sections below up to and including the choice of integration method specify things, data object creation and algorithm call create and compute things, and the remainder deals with the retrieval of estimation results and memory conservation.

Data entry

The methods below are used to enter data into Grumps. With Sources() one specifies where the data can be found and with Variables() which variables to use from those data sources. Note that there are two versions of the Variables method, where the main difference is the syntax. Use whichever one you prefer.

Grumps.Sources — Method

Sources( 
    T           = DefaultSourceTypes; 
    consumers   :: Any = nothing, 
    products    :: Any = nothing, 
    marketsizes :: Any = nothing, 
    draws       :: Any = nothing,
    user        :: Any = nothing
)

Creates a GrumpsSources object with source type entries of type T where the entries are provided in the optional parameters.

Grumps (potentially) uses four data sources: a data source for consumer-level data, one for product-level data, one for market size information, and one for demographic draws. Only the product-level data are required, but are by themselves insufficient. For instance, for BLP95 one needs information on products, market sizes, and demographics; for the Grumps estimator one needs all four types of data; for a multinomial logit both consumer and product information are needed. Not all data are needed for all markets. For instance, it is ok for some estimators for there to be consumer-level data in some markets but not others.

The T argument is mostly there to allow for future expansion, so the description below applies to the case in which T = DefaultSourceTypes.

By default, the entries can be nothing, a string, a DataFrame, or a SourceFileType. If an entry is nothing, it means that no such data is to be used. If an entry is a string then it is converted to a SourceFileCSV entry with comma delimiter where the string name is the file name. To use other source file types, create a SourceFileType first. A DataFrame can be passed, also. In all cases other than nothing, data will eventually be (converted to) a DataFrame and parsed from that.

The consumers variable specifies where consumer-level data can be found, the products variable is for the product-level data, marketsizes is for market sizes, and draws is for demographic draws; user has not been implemented yet.

Use the Variables() method to specify the way the data sources are formatted and the specification to estimate.

Grumps.Variables — Method

Variables( ; 
  market              :: Symbol = :market,
  choice              :: Symbol = :choice,
  interactions        :: Mat{Symbol} = [],
  randomcoefficients  :: Vec{Symbol} = [],
  outsidegood         :: String = "outsidegood",
  share               :: Symbol = :share,
  marketsize          :: Symbol = :N,
  regressors          :: Vec{Symbol} = [],
  instruments         :: Vec{Symbol} = [],
  dummies             :: Vec{Symbol} = [],
  nuisancedummy       :: Symbol = :none,
  microinstruments    :: Mat{Symbol} = [],
  user                :: Mat{Symbol} = []
    )

This method is used to specify regressors, instruments, random coefficients, interactions, variable labels, etcetera, from the sources you have specified in Sources(). It creates an object of type GrumpsVariables.

For instance, the option market specifies the column heading of the column containing the market descriptor (name). The same is true for all other arguments, except outsidegood which describes the spreadsheet entry that indicates the product is an outside good. The same label for the outside good should be used in all spreadsheets and all markets. Outside good entries should only be used in the consumer micro data and then only if there actually are consumers in the micro data choosing the outside good. All descriptors are case and space sensitive.

There is a separation between variables that go into the individual consumer utility and ones that only go into "mean utility". For instance, interactions tells Grumps which interaction terms to use and randomcoefficients which product level regressors are hit with a random coefficient. By contrast, regressors go into the mean utility component and are regressors in the "second stage" (where β is recovered). One can use the special symbol :constant to indicate a constant is to be used; the spreadsheet need not include a column with that heading.

Note that there are three ways that dummy variables can be entered as second stage regressors. The first is via regressors, in which case the onus is on the user to ensure that they have the correct numerical values. The second possibility is via the dummies argument. For each symbol passed via the dummies argument, Grumps will examine the corresponding column of the product data set (which can contain descriptive entries that need not be numerical) and turn it into dummy variables. If the coefficient on the dummies is of no interest then it is better to pass one via the nuisancedummy argument since it saves both computation time and memory. There can only be at most one categorical variable that can be converted to nuisance dummies, but there can be arbitrarily many categories. These dummies and nuisance dummies are automatically assumed to be exogenous and will be included in the instruments, also.

market refers to the variable containing the market indicator in all input datasets. Strings work best for the market indicators themselves, but it is not a requirement.

product refers to the variable containing the product indicator in the product dataset. Strings work best for the product indicators themselves, but it is not a requirement.

choice refers to the variable indicating the choice indicator in the consumer level datasets. Strings work best for the choice indicators themselves, but it is not a requirement.

interactions refers to the variables indicating consumer and product variable interactions (each row contains consumer variable, product variable)

randomcoefficients refers to the product level variables that have a random coefficient on them

outsidegood refers to the label used for the outside good

share refers to the label used for the product level share; these are shares where the denominator includes the outside good

marketsize refers to the size of the market (number of people)

regressors refers to the label used for the second stage regressors

instruments refers to the label used for the second stage instruments

dummies refers to discrete variables to be converted to second stage dummy regressors and instruments

nuisancedummy refers to at most one variable to be converted to a second stage dummy regressors and instrument whose coefficient value is of no interest

microinstruments refers to micro instruments, which are only relevant for gmm style procedures

user refers to a list of variables to be added to the consumer-product interactions using a user-specified procedure

Grumps.Variables — Method

Variables( 
    microspec           :: String,
    macrospec           :: String,
    dummyspec           :: String = "",
    nuisancedummyspec   :: String = "";
    market              :: String = "market",
    product             :: String = "product",
    outsidegood         :: String = "outsidegood",
    marketsize          :: String = "N",
    microinstruments    :: String = "",
    user                :: String = ""
    )

This method is used to specify regressors, instruments, random coefficients, interactions, variable labels, etcetera, from the sources you have specified in Sources(). It creates an object of type GrumpsVariables. There is another method by the same name that accomplishes much the same thing, but has a different user interface.

The method described here takes string arguments, including two mandatory ones, whereas the other method takes optional arguments only, mostly symbols, vectors of symbols, and matrices of symbols that can be passed in arbitrary order using keywords. The current method parses user input before calling the other method.

The option market specifies the column heading of the column containing the market descriptor (name). The same is true for all other arguments, except outsidegood which describes the spreadsheet entry that indicates the product is an outside good. The same label for the outside good should be used in all spreadsheets and all markets. Outside good entries should only be used in the consumer micro data and then only if there actually are consumers in the micro data choosing the outside good. All descriptors are case sensitive.

There is a separation between variables that go into the individual consumer utility and ones that only go into "mean utility". For instance, to specify what goes into individual consumer utility, one could specify

"choice = loginc * msrp + famsize * logfootprint + famsize * van + urban * truck + rc * suv + rc * truck + rc * van"

as the microspec argument to indicate that consumer choice is in the micro data set in a column headed "choice" and that there are four interaction terms and three random coefficients. The interaction terms have the consumer-level variable as the first factor and the product variable as the second argument. In this example the three random coefficients are on the suv, truck, and van variables and these variable names should correspond to the column headings in the product level data set. One can use constant to indicate a constant is used: there is no need to include a constant in one's data.

The macrospec argument takes the form

"share = constant + logmpg + loghp + logfootprint + msrp / constant, logmpg, loghp, logfootprint, logcurbweight, lagplcon"

where share are product-level market shares, everything between = and / represents regressors, and everything after / represents instruments; both regressors and instruments are for the product level moments portion. One can again use constant to indicate a constant is used, which need not be included in one's data.

Note that there are three ways that dummy variables can be entered as second stage regressors. The first is via macrospec, in which case the onus is on the user to ensure that they have the correct numerical values. The second possibility is via the dummyspec argument. For each variable passed via the dummyspec argument, Grumps will examine the corresponding column of the product data set (which can contain descriptive entries that need not be numerical) and turn it into dummy variables. If the coefficient on the dummies is of no interest then it is better to pass one via the nuisancedummyspec argument since it saves both computation time and memory. There can only be at most one categorical variable that can be converted to nuisance dummies, but there can be arbitrarily many categories. These dummies and nuisance dummies are automatically assumed to be exogenous and will be included in the instruments, also.

market refers to the variable containing the market indicator in all input datasets. Strings work best for the market indicators themselves, but it is not a requirement.

product refers to the variable containing the product indicator in the product dataset. Strings work best for the product indicators themselves, but it is not a requirement.

choice refers to the variable indicating the choice indicator in the consumer level datasets. Strings work best for the choice indicators themselves, but it is not a requirement.

interactions refers to the variables indicating consumer and product variable interactions (each row contains consumer variable, product variable)

randomcoefficients refers to the product level variables that have a random coefficient on them

outsidegood refers to the label used for the outside good

share refers to the label used for the product level share; these are shares where the denominator includes the outside good

marketsize refers to the size of the market (number of people)

regressors refers to the label used for the second stage regressors

instruments refers to the label used for the second stage instruments

dummies refers to discrete variables to be converted to second stage dummy regressors and instruments

nuisancedummy refers to at most one variable to be converted to a second stage dummy regressors and instrument whose coefficient value is of no interest

microinstruments refers to micro instruments, which are only relevant for gmm style procedures

user refers to a list of variables to be added to the consumer-product interactions using a user-specified procedure

Optimization options

The default optimization options are sensible, in which case this section can be skipped. But for those who want to play with tolerances and such, have at it.

There is one exception, however, and that exception pertains to using less memory. There is a separate section dedicated to that possibility, namely Memory conservation

Grumps.OptimizationOptions — Method

OptimizationOptions(; 
θopt = OptimOptionsθ(), 
δopt = OptimOptionsδ(), 
threads = GrumpsThreads(), 
memsave = false, 
maxrepeats = 4, 
probtype = :fast )

Sets the options used for numerical optimization. θopt is used for the external optimization routine, δopt for the internal one. These are both of type OptimOptions; see the OptimOptionsθ and OptimOptionsδ methods for elaboration. The memsave variable is set to false by default; turning it on will reduce memory consumption significantly, but will also slow down computation. The variable maxrepeats may disappear in the future.

Finally, there are two ways of computing choice probabilities: robust and fast, specified by passing :robust or :fast in probtype. Fast choice probabilities are the default for good reason.

Grumps.OptimOptionsθ — Method

OptimOptionsθ(; 
f_tol = 1.0e-8, 
g_tol = 1.0e-4, 
x_tol = 1.0e-5, 
iterations = 25, 
show_trace = true, 
store_trace = true, 
extended_trace = true )

Creates and returns an OptimOptions optimization options variable for the outer optimization algorithm, including the function value tolerance, the gradient tolerance, the solution tolerance, the maximum number of iterations, whether to show the trace, whether to store the trace, and whether to keep the extended trace. See the Optim package for details.

The current version of Grumps will largely ignore the trace-related parameters.

Grumps.OptimOptionsδ — Method

OptimOptionsδ( ; 
f_tol = 1.0e-8, 
g_tol = 1.0e-8, 
x_tol = 1.0e-6, 
iterations = 25, 
show_trace = false, 
store_trace = true, 
extended_trace = false )

Creates and returns an OptimOptions optimization options variable for the inner optimization algorithm, including the function value tolerance, the gradient tolerance, the solution tolerance, the maximum number of iterations, whether to show the trace, whether to store the trace, and whether to keep the extended trace. See the Optim package for details.

The current version of Grumps will largely ignore the trace-related parameters.

Grumps.GrumpsThreads — Method

GrumpsThreads(; 
    blas = 0, 
    markets = 0, 
    inner = 0 
    )

This sets the number of threads to be used subject to a number of caveats. blas refers to the number of BLAS threads, markets to the number of threads in loops over markets, and inner to the number of threads in inner loops. A value of zero forces the automatic selection of the number of threads.

Of these, inner is not currently used at all, market is only used in memsave mode, and blas is used. However, please note that the number of threads used by Grumps altogether is the number of threads passed in via the command line argument (i.e. via the -t switch), where that number does not include the number of BLAS threads set.

Data storage options

The default data storage options are sensible, but some space can be saved by tinkering with the settings. However, the only parameter that is worth changing is σ2, which is the variance of ξ, the product level error term. This is of no relevance for two-stage estimators like unpenalized mle.

Grumps.DataOptions — Method

DataOptions(; 
    micromode   = :Hog
    macromode   = :Ant
    balance     = :micro
    σ2          = 1.0
)

Specifies how Grumps should store its data and what it should store. The first three options are best left alone, unless you know what it is you're doing. The last option is the variance of ξ, i.e. the error variance in the product level moments.

Standard error options

stub

Grumps.StandardErrorOptions — Method

StandardErrorOptions(; θ = true, δ = true, β = true, setype = :homo )

Specifies which coefficients to create standard errors for and what type of standard errors to produce. Current choices are :homo (i.e. assuming homoskedasticity) and :hetero (heteroskedasticity-robust). Fancier options will be added at a future point in time.

Estimator choice

Grumps can compute quite a few estimators and one can specify which estimator to use by passing the return value of a call to Estimator to the optimization routine.

The easiest way to call Estimator is by passing it a string that describes what it is that you want to do. The following estimators are currently defined:

the full Grumps estimator

a less expensive alternative estimator with the same limit distribution as the full Grumps estimator

Grumps-style maximum likelihood, i.e Grumps without penalty

ditto, but imposing share constraints

GMM estimator that uses both micro and macro moments and uses quadrature instead of Monte Carlo draws in the micro moments. The micro moments are smart in that they condition on $z_{im}$ instead of integrating it out.

a mixed logit estimator

For a description of these estimators, see Estimator.

Grumps.Estimator — Method

Estimator( s :: String )

Creates and returns a GrumpsEstimator type. Grumps is reasonably good at figuring out what it is that you want, so e.g. Estimator( "maximum likelihood" ) gives you the unpenalized Grumps maximum likelihood estimator.

The estimators currently programmed include:

the full Grumps estimator

Grumps-style maximum likelihood, i.e Grumps without penalty

ditto, but imposing share constraints

GMM estimator that uses both micro and macro moments and uses quadrature instead of Monte Carlo draws in the micro moments. The micro moments are `smart' in that they condition on $z_{im}$ instead of integrating it out.

a mixed logit estimator

Grumps.Estimator — Method

Estimator( s :: Symbol )

Creates and returns a GrumpsEstimator type.

This is one method of specifying the estimator used. However, it is unforgiving in that the exact symbol used internally must be passed, so the Estimator( s :: String ) method is usually a better choice.

Possible choices include:

:pml the full Grumps maximum likelihood estimator

:vanilla the unpenalized Grumps maximum likelihood estimator

:shareconstraint the unpenalized Grumps maximum likelihood estimator with share constraints

:gmm GMM estimator that uses both micro and macro moments

:mixedlogit mixed logit maximum likelihood estimator

Grumps.Estimators — Method

Estimators( elaborate = false )

Prints a list of available estimators. The argument indicates whether a lot of features should be printed or few.

Choice of integration method (integrators)

Grumps uses separate integration methods for the micro and macro components. The default choices are simple with small numbers of nodes and draws. For micro, it is Hermitian quadrature, for macro it's Monte Carlo draws. One gets the defaults if the choices are omitted. The defaults chosen here are small in the sense that they emphasize speed / storage over accuracy. To change the number of nodes or draws, simply call BothIntegrators with as argument(s), whichever of the two you wish to change. For instance, integ = BothIntegrators( DefaultMicroIntegrator( 19 ) ) uses the default micro integrator with 19 nodes per dimension and the default macro integrator with the default number of draws.

The procedure is to create the integrators using a call to BothIntegrators with the desired integrators as arguments and then pass this in your call to Data.

Grumps.BothIntegrators — Method

BothIntegrators( microIntegrator :: MicroIntegrator{T}, macroIntegrator :: MacroIntegrator{T} )

Creates the type BothIntegrators containing both the indicated microIntegrator and macroIntegrator.

Either argument can be omitted. If both arguments are omitted then one can pass the floating point type T instead. If no floating point type is passed then a Float64 is assumed.

Grumps.DefaultMicroIntegrator — Method

DefaultMicroIntegrator( n :: Int, T :: Type )

Creates a basic quadrature Integrator using n nodes in each dimension. Type T can be omitted, in which case it is Float64.

Grumps.DefaultMacroIntegrator — Method

DefaultMacroIntegrator( n :: Int, T :: Type )

Creates a basic Monte Carlo Integrator using n draws. Type T can be omitted, in which case it is Float64.

Grumps.MSMMicroIntegrator — Method

MSMMicroIntegrator( n :: Int, T = F64 )

Creates a Monte Carlo integrator type for micro integration with GMM with smart moments. The optional type can be omitted.

Data object creation

The data stored in spreadsheets or other objects have to be converted into a form that Grumps understands. The call to Data achieves that. It takes as inputs the various choices made by the user and then creates an appropriate data object that is subsequently passed to the optimization call.

Grumps.Data — Method

Data( 
    e                   :: GrumpsEstimator,
    ss                  :: Sources,
    v                   :: Variables,
    integrators         :: GrumpsIntegrators = BothIntegrators(),
    T                   :: Type = F64,
    u                   :: UserEnhancement = DefaultUserEnhancement();
    options             :: DataOptions = GrumpsDataOptions(),
    threads             :: Int = 0
    )

Takes user inputs and converts them into an object that Grumps can understand. This is synonymous with GrumpsData(...).

Data takes the following arguments, of which the first three are mandatory:

e: estimator; see Estimator

ss: cata sources; see Sources

v: variables to be used; see Variables

o: optimization options to be used

integrators: see BothIntegrators, DefaultMicroIntegrator, and DefaultMacroIntegrator

T: floating point type; not heavily tested

u: not yet implemented

options: data options to be used, see DataOptions

threads: the number of parallel threads to be used in creating data

Algorithm call

Once all data structures have been put together, one can call the algorithm. This is straightforward.

Grumps.grumps! — Method

grumps!( 
    e       :: Estimator,
    d       :: Data{T},
    o       :: OptimizationOptions = OptimizationOptions(),
    θstart  :: StartingVector{T} = nothing,
    seo     :: StandardErrorOptions = StandardErrorOptions()
)

Conducts the optimization. You typically just want to set θstart to nothing, i.e. have a starting vector picked automatically.

Retrieving results

As noted above, Grumps will return its results in a GrumpsSolution variable that can be queried or saved as follows. You can also simply call one of the print or related functions on any of these objects.

Finally, you can call any of minimum, iterations, iteration_limit_reached, converged, f_converged, g_converged, x_converged, f_calls, g_calls, h_calls, f_trace, g_norm_trace, x_trace on a GrumpsSolution or a GrumpsConvergence object in the same way that you would query the return value in the Optim package, albeit that they are not in the namespace by default so use Grumps.converged instead of converged.

Grumps.getθ — Method

getθ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for θ that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getδ — Method

getδ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for δ that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getβ — Method

getβ( sol :: GrumpsSolution )

Returns a vector of GrumpsEstimate types for β that can be queried for results. See getcoef, getstde, gettstat, and getname.

Grumps.getcoef — Method

getcoef( e :: GrumpsEstimate )

Returns the estimated coefficient value.

Grumps.getstde — Method

getstde( e :: GrumpsEstimate )

Returns the standard error.

Grumps.gettstat — Method

gettstat( e :: GrumpsEstimate )

Returns the t statistic.

Grumps.getname — Method

getname( e :: GrumpsEstimate )

Returns the variable name.

Grumps.getθcoef — Method

getθcoef( sol :: GrumpsSolution )

Returns a vector of θ coefficients

Grumps.getδcoef — Method

getδcoef( sol :: GrumpsSolution )

Returns a vector of δ coefficients

Grumps.getβcoef — Method

getβcoef( sol :: GrumpsSolution )

Returns a vector of β coefficients

Grumps.Save — Method

Save( fn, mt, sol :: GrumpsSolution; keywords... )

Saves the solution stored in sol to a file with filename fn which has mime type mt.

There are several keywords that are described below, some of which will be ignored for some mime types. Allowed mime types are text/plain, text/csv, and text/tex.

Keyword	Description	Default

colsep	column separator	","

adorned	make output pretty?	true

printθ	print θ results?	true

printβ	print β results?	true

printδ	print δ results?	false

printconvergence	convergence stats?	true

Grumps.Save — Method

Save( fn, sol :: GrumpsSolution; keywords... )

The same as the form of Save with prespecified mime type except that the mime type is now inferred from the file extension. The keywords also have the same meaning, namely...

There are several keywords that are described below, some of which will be ignored for some mime types. Allowed mime types are text/plain, text/csv, and text/tex.

Keyword	Description	Default

colsep	column separator	","

adorned	make output pretty?	true

printθ	print θ results?	true

printβ	print β results?	true

printδ	print δ results?	false

printconvergence	convergence stats?	true

Base.show — Method

show( io :: IO, e :: GrumpsEstimate{T}, s :: String = ""; keywords...)

Show a GrumpsEstimate object on io; the argument s indicates which parameter family (θ,β,δ) the estimate belongs to. The optional keywords are described in the table below.

Keyword	Description	Default

adorned	make output pretty?	true

printstde	print standard errors?	true

printtstat	print t statistics?	true

Base.show — Method

show( io :: IO, est :: Vector{ GrumpsEstimate{T} }, s :: String = ""; keywords... )

Shows a vector of estimates on io using the string s (typically one of θ,β,δ). The command takes the following optional keywords.

Keyword	Description	Default

adorned	make output pretty?	true

printstde	print standard errors?	true

printtstat	print t statistics?	true

header	descriptive header?	false

Base.show — Method

show( io :: IO, convergence :: GrumpsConvergence{T}; keywords...)

Shows the contents of convergence, where the flags indicated what should be printed and how, as indicated in the following table.

Keyword	Description	Default

adorned	make output pretty?	true

header	descriptive header?	false

Base.show — Method

show( io :: IO, sol :: GrumpsSolution{T}; keywords... )

Shows the contents of sol, where the keywords indicate what should be printed and how, as described in the table below.

Keyword	Description	Default

adorned	make output pretty?	true

printθ	print θ results?	true

printβ	print β results?	true

printδ	print δ results?	false

printconvergence	convergence stats?	true

Base.show — Method

show( io :: IO, mt :: MIME{Symbol("text/tex")}, sol :: GrumpsSolution; keywords... )

This is the same as Save() except that the contents are spit out on io (which could be stdout or an already opened file).

Base.show — Method

show( io :: IO, mt :: MIME{Symbol("text/csv")}, sol :: GrumpsSolution; keywords... )

This is the same as Save() except that the contents are spit out on io (which could be stdout or an already opened file).

Memory conservation

stub; this section to be completed

By default, Grumps loads all data and then creates space for all markets for things like choice probabilities, objective functions and their derivatives, intermediate objects, etcetera. This saves computation time, but eats memory, especially as the number of random coefficients increases.

To conserve memory, one can set memsave in OptimizationOptions() to true. What this does is that it shares space for choice probabilities and related objects across a number of markets. For instance, if there are ten markets and the number of market threads in OptimizationOptions() is set to two then the space for choice probabilities is shared across five markets. These choices will have no effect if the number of market threads is no less than the number of markets. The downside of doing this is that it slows down computation since choice probabilities need to be recomputed. This is especially true for estimators that use the penalty term.

There are less impactful ways of reducing memory usage, such as choosing the option :Ant for the micro data, also. *** not yet implemented ***